+++ packages/core-data-ui/src/components/entities-saved-states/entity-record-item.js
@@ -10,23 +10,16 @@ import { decodeEntities } from '@wordpress/html-entities';
 /**
  * Internal dependencies
  */
-import { store as editorStore } from '../../store';
-import { unlock } from '../../lock-unlock';
 import { getTemplateInfo } from '../../utils/get-template-info';
 
 export default function EntityRecordItem( { record, checked, onChange } ) {
 	const { name, kind, title, key } = record;
 
 	// Handle templates that might use default descriptive titles.
-	const { entityRecordTitle, hasPostMetaChanges } = useSelect(
+	const entityRecordTitle = useSelect(
 		( select ) => {
 			if ( 'postType' !== kind || 'wp_template' !== name ) {
-				return {
-					entityRecordTitle: title,
-					hasPostMetaChanges: unlock(
-						select( editorStore )
-					).hasPostMetaChanges( name, key ),
-				};
+				return title;
 			}
 
 			const template = select( coreStore ).getEditedEntityRecord(
@@ -38,36 +31,24 @@ export default function EntityRecordItem( { record, checked, onChange } ) {
 			const { default_template_types: templateTypes = [] } =
 				select( coreStore ).getCurrentTheme() ?? {};
 
-			return {
-				entityRecordTitle: getTemplateInfo( {
-					template,
-					templateTypes,
-				} ).title,
-				hasPostMetaChanges: unlock(
-					select( editorStore )
-				).hasPostMetaChanges( name, key ),
-			};
+			return getTemplateInfo( {
+				template,
+				templateTypes,
+			} ).title;
 		},
 		[ name, kind, title, key ]
 	);
 
 	return (
-		<>
-			<PanelRow>
-				<CheckboxControl
-					label={
-						decodeEntities( entityRecordTitle ) || __( 'Untitled' )
-					}
-					checked={ checked }
-					onChange={ onChange }
-					className="entities-saved-states__change-control"
-				/>
-			</PanelRow>
-			{ hasPostMetaChanges && (
-				<ul className="entities-saved-states__changes">
-					<li>{ __( 'Post Meta.' ) }</li>
-				</ul>
-			) }
-		</>
+		<PanelRow>
+			<CheckboxControl
+				label={
+					decodeEntities( entityRecordTitle ) || __( 'Untitled' )
+				}
+				checked={ checked }
+				onChange={ onChange }
+				className="entities-saved-states__change-control"
+			/>
+		</PanelRow>
 	);
 }