+++ packages/core-data-ui/src/components/entities-saved-states/index.js
@@ -18,13 +18,13 @@ import {
 	useInstanceId,
 } from '@wordpress/compose';
 import { useDispatch } from '@wordpress/data';
+import { store as coreStore } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
  */
 import EntityTypeList from './entity-type-list';
 import { useIsDirty } from './hooks/use-is-dirty';
-import { store as editorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 
 /**
@@ -87,7 +87,7 @@ export function EntitiesSavedStatesExtensible( {
 	successNoticeContent,
 } ) {
 	const saveButtonRef = useRef();
-	const { saveDirtyEntities } = unlock( useDispatch( editorStore ) );
+	const { saveDirtyEntities } = unlock( useDispatch( coreStore ) );
 	// To group entities by type.
 	const partitionedSavables = dirtyEntityRecords.reduce( ( acc, record ) => {
 		const { name } = record;