+++ packages/core-data-ui/src/index.js
@@ -1,0 +1,4 @@
+export { default as EntitiesSavedStates } from './components/entities-saved-states';
+export { useIsDirty as useEntitiesSavedStatesIsDirty } from './components/entities-saved-states/hooks/use-is-dirty';
+
+export { privateApis } from './private-apis';